Moved feature_home to app module as it does not make sense to be a dynamic feature
