+++ app/src/main/res/drawable/avd_theme_day_to_night.xml
