+++ app/src/main/res/drawable/avd_theme_night_to_day.xml
