+++ app/src/main/res/drawable/ic_favorites.xml
