+++ app/src/main/res/drawable/ic_theme_day.xml
