+++ app/src/main/res/drawable/sl_navigation.xml
