+++ app/src/main/res/values/colors_home.xml
