+++ app/src/main/res/values/strings_home.xml
