--- feature_home/src/main/AndroidManifest.xml
@@ -1,19 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:dist="http://schemas.android.com/apk/distribution"
-    xmlns:tools="http://schemas.android.com/tools"
-    package="com.jpaya.dynamicfeatures.home">
-
-    <dist:module
-        dist:instant="true"
-        dist:title="@string/feature_home">
-        <dist:delivery>
-            <dist:install-time />
-        </dist:delivery>
-        <dist:fusing dist:include="false" />
-    </dist:module>
-
-    <application
-        android:supportsRtl="true"
-        tools:ignore="AllowBackup,MissingApplicationIcon" />
-
-</manifest>
