--- feature_home/src/main/kotlin/com/jpaya/dynamicfeatures/home/ui/di/HomeModule.kt
@@ -1,50 +1,0 @@
-/*
- * Copyright 2020 Jose Maria Payá Castillo
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.jpaya.dynamicfeatures.home.ui.di
-
-import androidx.annotation.VisibleForTesting
-import androidx.annotation.VisibleForTesting.PRIVATE
-import com.jpaya.base.ui.extensions.viewModel
-import com.jpaya.dynamicfeatures.home.ui.HomeFragment
-import com.jpaya.dynamicfeatures.home.ui.HomeViewModel
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.android.components.FragmentComponent
-
-/**
- * Class that provides home-related dependencies to the hilt dependency graph [FragmentComponent].
- * All dependencies provided by this class will be considered as fragment-level properties.
- */
-@Module
-@InstallIn(FragmentComponent::class)
-class HomeModule(
-    @VisibleForTesting(otherwise = PRIVATE)
-    val fragment: HomeFragment
-) {
-
-    /**
-     * Create a provider method binding for [HomeViewModel].
-     *
-     * @return Instance of view model.
-     * @see Provides
-     */
-    @Provides
-    fun providesHomeViewModel() = fragment.viewModel {
-        HomeViewModel()
-    }
-}
