+++ settings.gradle.kts
@@ -17,7 +17,6 @@
 include(
     ":app",
     ":base-android",
-    ":feature_home",
     ":feature_abbreviations",
     ":libraries:test_utils"
 )
